% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file_shapefile.R
\name{download_file_shapefile}
\alias{download_file_shapefile}
\title{Download file shapefile}
\usage{
download_file_shapefile(index)
}
\arguments{
\item{index}{Variable define type of shapefile}
}
\description{
Function return files format shapefile
}
\examples{
\dontrun{

library(geobage)

  c <- download_file_shapefile()

}
}
\seealso{
Other general area functions: 
\code{\link{read_biome_pampa}()},
\code{\link{read_census_tract_bg}()},
\code{\link{read_conservation_unit_bg}()},
\code{\link{read_health_facilities_bg}()},
\code{\link{read_immediate_region_bg}()},
\code{\link{read_info_bg}()},
\code{\link{read_intermediate_region_bg}()},
\code{\link{read_meso_region_bg}()},
\code{\link{read_micro_region_bg}()},
\code{\link{read_municipal_seat_bg}()},
\code{\link{read_public_rural_property_bg}()},
\code{\link{read_shape_bg}()},
\code{\link{read_urban_area_bg}()}
}
\concept{general area functions}
