% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_area_risk_bg.R
\name{read_area_risk_bg}
\alias{read_area_risk_bg}
\title{Area of risk}
\usage{
read_area_risk_bg()
}
\description{
Function returns the area of risk inserted in municipality of Bage
}
\examples{
\dontrun{

library(geobage)

  c <- read_area_risk_bg()

}
}
\seealso{
Other general area functions: 
\code{\link{download_file_shapefile}()},
\code{\link{read_biome_pampa}()},
\code{\link{read_census_tract_bg}()},
\code{\link{read_conservation_unit_bg}()},
\code{\link{read_drainage_network_bg}()},
\code{\link{read_health_facilities_bg}()},
\code{\link{read_immediate_region_bg}()},
\code{\link{read_info_bg}()},
\code{\link{read_intermediate_region_bg}()},
\code{\link{read_macrozones_bg}()},
\code{\link{read_meso_region_bg}()},
\code{\link{read_micro_region_bg}()},
\code{\link{read_municipal_seat_bg}()},
\code{\link{read_neighborhoods_bg}()},
\code{\link{read_public_rural_property_bg}()},
\code{\link{read_shape_bg}()},
\code{\link{read_urban_area_bg}()}
}
\concept{general area functions}
