% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biome_pampa.R
\name{read_biome_pampa}
\alias{read_biome_pampa}
\title{Biome Pampa}
\usage{
read_biome_pampa()
}
\description{
Function returns the Pampa Biome, biome where the municipality of Bagé is inserted
}
\examples{
\dontrun{

library(geobage)

  c <- read_biome_pampa()

}
}
\seealso{
Other general area functions: 
\code{\link{read_census_tract_bg}()},
\code{\link{read_conservation_unit_bg}()},
\code{\link{read_health_facilities_bg}()},
\code{\link{read_immediate_region_bg}()},
\code{\link{read_info_bg}()},
\code{\link{read_intermediate_region_bg}()},
\code{\link{read_meso_region_bg}()},
\code{\link{read_micro_region_bg}()},
\code{\link{read_municipal_seat_bg}()},
\code{\link{read_shape_bg}()},
\code{\link{read_urban_area_bg}()}
}
\concept{general area functions}
