% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_immediate_region_bg.R
\name{read_immediate_region_bg}
\alias{read_immediate_region_bg}
\title{Areas immediate the municipality of Bagé}
\usage{
read_immediate_region_bg()
}
\description{
Function returns the intermediate region in which the municipality of Bagé is inserted. Intermediate Geographic Areas are part of the geographical division of Brazil created in 2017 by IBGE. These regions were created to replace the "Micro Regions" division. Data on scale 1: 250,000, using the geodetic reference system "SIRGAS2000" and CRS (4674)
}
\examples{
\dontrun{

library(geobage)

  c <- read_immediate_region_bg()

}
}
\seealso{
Other general area functions: 
\code{\link{read_biome_pampa}()},
\code{\link{read_census_tract_bg}()},
\code{\link{read_conservation_unit_bg}()},
\code{\link{read_health_facilities_bg}()},
\code{\link{read_info_bg}()},
\code{\link{read_intermediate_region_bg}()},
\code{\link{read_meso_region_bg}()},
\code{\link{read_micro_region_bg}()},
\code{\link{read_municipal_seat_bg}()},
\code{\link{read_shape_bg}()},
\code{\link{read_urban_area_bg}()}
}
\concept{general area functions}
