% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_info_bg.R
\name{read_info_bg}
\alias{read_info_bg}
\title{Information about the municipality of Bagé}
\usage{
read_info_bg()
}
\description{
Function returns various information about the municipality of Bagé, such as name and codes of corresponding, meso, micro, intermediate and immediate regions of the municipality, among others
}
\examples{
\dontrun{

library(geobage)

  c <- read_info_bg()

}
}
\seealso{
Other general area functions: 
\code{\link{read_biome_pampa}()},
\code{\link{read_census_tract_bg}()},
\code{\link{read_conservation_unit_bg}()},
\code{\link{read_health_facilities_bg}()},
\code{\link{read_immediate_region_bg}()},
\code{\link{read_intermediate_region_bg}()},
\code{\link{read_meso_region_bg}()},
\code{\link{read_micro_region_bg}()},
\code{\link{read_municipal_seat_bg}()},
\code{\link{read_shape_bg}()},
\code{\link{read_urban_area_bg}()}
}
\concept{general area functions}
