% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_municipal_seat_bg.R
\name{read_municipal_seat_bg}
\alias{read_municipal_seat_bg}
\title{Municipality seat of Bagé}
\usage{
read_municipal_seat_bg()
}
\description{
Function returns the headquarters of the municipality of Bagé
}
\examples{
\dontrun{

library(geobage)

  c <- read_municipal_seat_bg()

}
}
\seealso{
Other general area functions: 
\code{\link{read_biome_pampa}()},
\code{\link{read_census_tract_bg}()},
\code{\link{read_conservation_unit_bg}()},
\code{\link{read_health_facilities_bg}()},
\code{\link{read_immediate_region_bg}()},
\code{\link{read_info_bg}()},
\code{\link{read_intermediate_region_bg}()},
\code{\link{read_meso_region_bg}()},
\code{\link{read_micro_region_bg}()},
\code{\link{read_shape_bg}()},
\code{\link{read_urban_area_bg}()}
}
\concept{general area functions}
