% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_public_rural_property_bg.R
\name{read_public_rural_property_bg}
\alias{read_public_rural_property_bg}
\title{Public rural property}
\usage{
read_public_rural_property_bg()
}
\description{
Function returns the Public rural property inserted in municipality of Bage
}
\examples{
\dontrun{

library(geobage)

  c <- read_public_rural_property_bg()

}
}
\seealso{
Other general area functions: 
\code{\link{download_file_shapefile}()},
\code{\link{read_biome_pampa}()},
\code{\link{read_census_tract_bg}()},
\code{\link{read_conservation_unit_bg}()},
\code{\link{read_health_facilities_bg}()},
\code{\link{read_immediate_region_bg}()},
\code{\link{read_info_bg}()},
\code{\link{read_intermediate_region_bg}()},
\code{\link{read_meso_region_bg}()},
\code{\link{read_micro_region_bg}()},
\code{\link{read_municipal_seat_bg}()},
\code{\link{read_shape_bg}()},
\code{\link{read_urban_area_bg}()}
}
\concept{general area functions}
