% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_urban_area_bg.R
\name{read_urban_area_bg}
\alias{read_urban_area_bg}
\title{Urban footprint}
\usage{
read_urban_area_bg()
}
\description{
Function returns to urban area of Bagé municipality
}
\examples{
\dontrun{

library(geobage)

  c <- read_urban_area_bg()

}
}
\seealso{
Other general area functions: 
\code{\link{read_biome_pampa}()},
\code{\link{read_census_tract_bg}()},
\code{\link{read_conservation_unit_bg}()},
\code{\link{read_health_facilities_bg}()},
\code{\link{read_immediate_region_bg}()},
\code{\link{read_info_bg}()},
\code{\link{read_intermediate_region_bg}()},
\code{\link{read_meso_region_bg}()},
\code{\link{read_micro_region_bg}()},
\code{\link{read_municipal_seat_bg}()},
\code{\link{read_shape_bg}()}
}
\concept{general area functions}
